Preliminary patch for test_manydbs for the condition variable wakeup threshold.
This makes the expected wakeup counts/ratio #defines so they can be
controlled on a per-target basis. Include the thresholds I need to
make the test run reliably on current NetBSD, where there's a bug
causing extra wakeups.

XXX: It might be better to rearrange the test to check the number of
XXX: wakeups _issued_ as well, or maybe instead. But that's a bigger deal.