First part of fix for performance issue with RTS on large RLE cells.
When iterating the recnos in an RLE cell, use a history store cursor
to visit only the keys that have history store content and skip the
others. Remember a representative of the others that we can use to
figure out what to do with the data in the cell itself, if anything.

For now, instead of using the representative, use a hacked up list of
skipped ranges and process these as well afterwards. The change above
is big enough for the moment; the code to do the second step properly
will come in a separate installment.

There is one thing this could do that it so far intentionally doesn't:
skip the complicated iterator code for cells with RLE 1, or maybe some
larger threshold. For now, that just makes the code harder to debug.