Draft of second part for fix to performance issue with RTS on large RLE cells.
Add __rollback_col_rle_modify, which is a special case version of
__rollback_col_modify for RLE cells with no extant history entries. It
works by rewriting the cell in place, which is gross but in the
absence of any kind of range update record is the best available way
to deal with a large RLE cell in bulk. The code aggressively asserts
that it's only handling the things it thinks it ought to be able to:
deleting the cell or clearing the value's end time.

Pass a flag through __rollback_ondisk_fixup_key and
__rollback_abort_ondisk_kv, and use it to (a) assert that only the
expected things happen and (b) call the new __rollback_col_rle_modify
instead of __rollback_col_modify.

In __rollback_abort_col_var, pass the new flag as false, except after
iterating through all the keys with history store entries, choose the
first key without as a representative and call down once with the new
flag as true on that representative. Because there are no history
store entries, the __rollback_col_rle_modify rewrite performs the
necessary update on all keys in the cell simultaneously, and for keys
that had history the cell change happens under the updates already
posted so they remain correct.

In __rollback_abort_col_row, just pass the new flag as false.

Does not work yet, though. Currently:
   - will assert false if it sees an overflow value, because that code
     hasn't been written yet;
   - the skipped-range hacks from the previous commit/first part are
     retained for now, because they're still useful for debugging;
   - the iteration over extant history store keys to skip over keys in
     the cell that have none does not actually work and I'm still
     trying to figure out why;
   - consequently the cell-rewriting code mostly hasn't been tested
     and is probably wrong.