Take a different approach for WT-8747.
Instead of prohibiting reads of nondurable data, track the durable
timestamps of data read by a transaction and prohibit committing with
a durable timestamp (or commit timestamp for non-prepared
transactions) earlier than that.

Keep the test changes; most of them are about only reading nondurable
data when we explicitly mean to. It seems best to avoid doing that
otherwise.