Inhibit reverse splits when moving up from the leftmost child in VLCS.
This avoids nuking that slot in the parent internal page and creating
an unfillable namespace gap.

This (not anything in eviction) was the missing piece needed to make
this approach work.

Note that the reverse split is only inhibited when it's the leftmost
child that would be removed. If the next child gets removed, the
reverse split will happen then, and that case is correctly accounted
for. Otherwise, eventually eviction will take care of the page safely.